feat: Add SizedEntryAttribute for variable-length array serialization
- Introduced SizedEntryAttribute to specify per-entry byte size in wire headers for collections.
- Updated BinaryPacketSerializer to handle SizedEntryAttribute during serialization and deserialization.
- Modified AbilityInfoEntity to change CashCost from long? to ushort?.
- Adjusted AbilityDataProvider, AbilityListResponse, and CareerAbilityResponse to accommodate changes in CashCost and MinimumRank.
- Refactored AbilityResolver to streamline mastery ability resolution logic.
- Updated PlayerInitPipeline to remove obsolete Count property and integrate MinimumRank and CashCost.
- Enhanced tests for BinaryPacketSerializer to validate SizedEntry functionality with various scenarios including custom widths and little-endian formats.
- Refined PlayerInitPipelineTests to verify new properties in career abilities.

+++ src/WorldServerV2/Data/Entities/AbilityInfoEntity.cs
@@ -37,7 +37,7 @@ public sealed class AbilityInfoEntity
     public short? Category { get; set; }
     public int? Flags { get; set; }
     public short? PointCost { get; set; }
-    public long? CashCost { get; set; }
+    public ushort? CashCost { get; set; }
     public int? StealthInteraction { get; set; }
     public int? AiRange { get; set; }
     public int? IgnoreCooldownReduction { get; set; }
